Bug Patch
A temporary fix for the bug whereby Pinky causes the game to crash by being in a state they're not supposed to be in. The patch simply involves telling Pinky to go towards the center of the ghost house. I still have yet to see if this will break the game, but at the very least, it'll stop it from crashing.